refactor: move WriteBuffer units to src/WriteBuffer subdirectory
