+++ src/WriteBuffer/WriteBuffer.cpp
@@ -1,0 +1,1 @@
+#include "WriteBuffer.hpp"
