+++ src/WriteBuffer/WriteBuffer.hpp
